separate out some functionality to make code better structured
