+++ pytimer/pytimer.py
@@ -14,34 +14,11 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from math import sqrt
-import numpy as np
 from time import perf_counter
 from typing import Union, Tuple, List, TextIO, Dict, Set
 from functools import wraps
-import logging
 from sys import stdout
-
-from sklearn.linear_model import LinearRegression
-from sklearn.preprocessing import PolynomialFeatures
-from scipy.optimize import curve_fit
-
-
-class LoggingIO:
-    """
-    Provides a basic wrapper around logging.info to allow it to act like a file-like object to be used
-    as an output stream in some of the functions below.
-    """
-    @staticmethod
-    def write(message: str):
-        """
-        Write a message to logging.info. If the message ends with a newline, remove it, as logging.info adds that itself
-        :param message: the string message to log
-        """
-        if message[-1] == "\n":
-            logging.info(message[:-1])
-        else:
-            logging.info(message)
+from .data_structures import Run, Split
 
 
 class BaseTimer:
@@ -355,290 +332,6 @@ class StaticTimer(BaseTimer):
             else:
                 # Tuple[any, List[float]]
                 return value, [StaticTimer._convert_time(time, output_unit) for time in run_totals]
-
-
-class BestFitBase:
-    """
-    An abstract class to be used as a template for best fit curves. The process is to first calculate the curve using
-    calculated points and then check how accurate the curve is.
-    """
-    @staticmethod
-    def calculate_curve(points: List[Tuple[Tuple[List[int], Dict[str, int]], float]]) -> dict:
-        """
-        Take a list of tuples, each containing the arguments and the time value, and determines the parameters for this
-        type of curve. All arguments must be integers.
-        :param points: each entry is ((tuple of positional arguments, dict of keyword arguments), measured time)
-        :return: a dict of the parameters of the curve
-        """
-        pass
-
-    @staticmethod
-    def calculate_point(arguments: Tuple[List[int], Dict[str, int]], parameters: dict) -> float:
-        """
-        Take a tuple of arguments and calculate what the time should be for those arguments with the given parameters
-        :param arguments: positional and keyword arguments. All values must be ints.
-        :param parameters: the parameters of the calculated curve
-        :return: the time value for the given arguments and parameters
-        """
-        pass
-
-    @staticmethod
-    def flatten_args_separate_points(points: List[Tuple[Tuple[List[int], Dict[str, int]], float]]
-                                     ) -> Tuple[List[List[int]], List[float]]:
-        """
-        Take a list of points and separate them two lists, one containing a list of all the args and kwargs flattened,
-        the other containing the measured times corresponding to that list of arguments. Flattening the kwargs requires
-        dict.values() to be stable.
-        """
-        flattened_args = []
-        matching_points = []
-        for all_args, y in points:  # flatten args and kwargs into a single list. Requires dicts to be stable
-            matching_points.append(y)
-
-            args = [arg for arg in all_args[0]]
-            args.extend(all_args[1].values())  # relies on keys(), and values() being stable
-            flattened_args.append(args)
-
-        return flattened_args, matching_points
-
-    @staticmethod
-    def poll(points: List[Tuple[Tuple[List[int], Dict[str, int]], float]]) -> bool:
-        """
-        Determine if this best fit method will work with the given data
-        """
-        return True
-
-
-class BestFitExponential(BestFitBase):
-    @staticmethod
-    def calculate_curve(points):
-        """
-        Use scipy.optimize.curve_fit with a*e^(b*x) to find a and b for this exponential curve
-        """
-        flattened_args, matching_points = BestFitBase.flatten_args_separate_points(points)
-        values = curve_fit(lambda x, a, b: a*np.exp(b*x), [val[0] for val in flattened_args], matching_points,
-                           p0=(0.000001, 0.000001))  # set default params low as measured times can be very short
-
-        return {"a": values[0][0], "b": values[0][1]}
-
-    @staticmethod
-    def calculate_point(arguments, parameters):
-        x = arguments[0][0] if arguments[0] else list(arguments[1].values())[0]  # get arg or kwarg
-        return parameters["a"] * np.exp(parameters["b"]*x)
-
-    @staticmethod
-    def poll(points):
-        """
-        There can only be one argument for a exponential curve, this guarantees that is the case
-        """
-        for args, _ in points:
-            if len(args[0]) + len(args[1]) != 1:
-                return False
-
-        return True
-
-
-class BestFitLinear(BestFitBase):
-    @staticmethod
-    def calculate_curve(points):
-        """
-        Use sklearn.linear_model.LinearRegression to determine the variable coefficients and y-intercept
-        """
-        flattened_args, matching_points = BestFitBase.flatten_args_separate_points(points)
-
-        model = LinearRegression()
-        model.fit(flattened_args, matching_points)
-
-        params = {"b": model.intercept_}
-        i = 0
-        for _ in range(len(points[0][0][0])):
-            params["x{}".format(i)] = model.coef_[i]
-            i += 1
-
-        for key in points[0][0][1]:
-            params[key] = model.coef_[i]
-            i += 1
-
-        return params
-
-    @staticmethod
-    def calculate_point(arguments, parameters):
-        """
-
-        """
-        value = parameters["b"]
-
-        for i in range(len(arguments[0])):
-            value += arguments[0][i] * parameters["x{}".format(i)]
-
-        for key in arguments[1]:
-            value += arguments[1][key] * parameters[key]
-
-        return value
-
-
-class BestFitLogarithmic(BestFitBase):
-    @staticmethod
-    def calculate_curve(points):
-        """
-        Use scipy.optimize.curve_fit with a + b*log(x) to find a and b for this logarithmic curve
-        """
-        flattened_args, matching_points = BestFitBase.flatten_args_separate_points(points)
-        values = curve_fit(lambda x, a, b: a + b*np.log(x), [val[0] for val in flattened_args], matching_points)
-
-        return {"a": values[0][0], "b": values[0][1]}
-
-    @staticmethod
-    def calculate_point(arguments, parameters):
-        x = arguments[0][0] if arguments[0] else list(arguments[1].values())[0]  # get arg or kwarg
-        return parameters["a"] + parameters["b"]*np.log(x)
-
-    @staticmethod
-    def poll(points):
-        """
-        There can only be one argument for a logarithmic curve, this guarantees that is the case
-        """
-        for args, _ in points:
-            if len(args[0]) + len(args[1]) != 1:
-                return False
-
-        return True
-
-
-class BestFitPolynomial(BestFitBase):
-    @staticmethod
-    def calculate_curve(points):
-        """
-        Use sklearn.linear_model.LinearRegression to determine the variable coefficients and y-intercept
-        """
-        flattened_args, matching_points = BestFitBase.flatten_args_separate_points(points)
-
-        poly_model = PolynomialFeatures(degree=2)
-        values = poly_model.fit_transform(flattened_args)
-
-        values_model = LinearRegression()
-        values_model.fit(values, matching_points)
-
-        params = {"b": values_model.intercept_}
-        for i in range(len(values_model.coef_)):
-            params["x^{}".format(i)] = values_model.coef_[i]
-
-        return params
-
-    @staticmethod
-    def calculate_point(arguments, parameters):
-        """
-
-        """
-        flattened_args = arguments[0][:]
-        flattened_args.extend(arguments[1].values())
-
-        poly_model = PolynomialFeatures(degree=2)
-        adjusted = poly_model.fit_transform([flattened_args])[0]
-
-        value = parameters["b"]
-        for i in range(len(adjusted)):
-            value += parameters["x^{}".format(i)] * adjusted[i]
-
-        return value
-
-
-class Run:
-    def __init__(self, label: str, time: float, runs: int, iterations_per_run: int, args: tuple=(), kwargs: dict=()):
-        self.label: str = label
-        self.time: float = time
-        self.runs = runs
-        self.iterations_per_run = iterations_per_run
-        self.args: tuple = args
-        self.kwargs: dict = kwargs if kwargs else {}
-
-
-class Split:
-    best_fit_curves = {"Exponential": BestFitExponential, "Linear": BestFitLinear, "Logarithmic": BestFitLogarithmic,
-                       "Polynomial": BestFitPolynomial}
-
-    def __init__(self, label: str="Split"):
-        self.runs: List[Run] = []
-        self.label = label
-
-    def add_run(self, run: Run):
-        self.runs.append(run)
-
-    def average(self) -> float:
-        if self.runs:
-            return sum(i.time for i in self.runs) / len(self.runs)
-        else:
-            return 0
-
-    def determine_best_fit(self, curve_type: str=any, exclude_args: Set[int]=(), exclude_kwargs: set=(),
-                           transformers: Dict[Union[int, str], callable]=()) -> Union[None, Tuple[str, dict]]:
-        """
-        Determine the best fit curve for the runs contained in this split.
-        :return: A tuple of a string name for the best fit curve and a dict of the parameters for that curve
-        """
-        points = []
-        for run in self.runs:
-            if not run.args and not run.kwargs:
-                raise RuntimeWarning("Arguments must have been logged to determine a best fit curve")
-
-            # TRANSFORM ARGUMENTS
-            new_args, new_kwargs = run.args[:], dict(run.kwargs)
-            if transformers:
-                for i in range(len(new_args)):
-                    if i in transformers:
-                        new_args[i] = transformers[i](new_args[i])
-
-                for key in new_kwargs:
-                    if key in transformers:
-                        new_kwargs[key] = transformers[key](new_kwargs[key])
-
-            # EXCLUDE ARGUMENTS
-            new_args = [new_args[i] for i in range(len(new_args)) if i not in exclude_args]
-
-            for key in exclude_kwargs:
-                del new_kwargs[key]
-
-            points.append(((new_args, new_kwargs), run.time))
-
-        if curve_type is any:
-            best: Tuple[float, str, dict] = None  # tuple of distance, name, and parameters
-            for bfc_name, bfc in self.best_fit_curves.items():
-                params = bfc.calculate_curve(points)
-
-                # DISTANCE
-                distance = 0
-                for point in points:
-                    distance += abs(point[1] - bfc.calculate_point(point[0], params))
-
-                if best is None or distance < best[0]:
-                    best = (distance, bfc_name, params)
-
-            return best[1], best[2]
-        else:
-            if curve_type in self.best_fit_curves:
-                return curve_type, self.best_fit_curves[curve_type].calculate_curve(points)
-            else:
-                raise RuntimeWarning("{} is an invalid curve type. Must be in [{}]".format(
-                    curve_type, ", ".join(self.best_fit_curves.keys())
-                ))
-
-    def standard_deviation(self) -> float:
-        """
-        Calculate the standard deviation of the runs contained within this split where SD is:
-        sqrt(sum (x - x_bar)**2 / n)
-        :return: the standard deviation
-        """
-        if self.runs:
-            avg = self.average()
-            return sqrt(sum((run.time - avg) ** 2 for run in self.runs) / len(self.runs))
-        else:
-            return 0
-
-    def variance(self) -> float:
-        if self.runs:
-            return self.standard_deviation() ** 2
-        else:
-            return 0
 
 
 class Timer(BaseTimer):
